Refactor auth logic and add password change feature
Authentication logic has been moved from UserService to a new AuthService and AuthServiceImpl, with AuthController updated to use AuthService. Added password change functionality to UserController and UserService, including PasswordChangeRequest DTO and implementation in UserServiceImpl. LoginResponse now returns UserResponse instead of UserRequest.
