+++ src/main/java/com/example/enotes/controller/AuthController.java
@@ -3,7 +3,7 @@ package com.example.enotes.controller;
 import com.example.enotes.dto.LoginRequest;
 import com.example.enotes.dto.LoginResponse;
 import com.example.enotes.dto.UserRequest;
-import com.example.enotes.service.UserService;
+import com.example.enotes.service.AuthService;
 import com.example.enotes.util.CommonUtil;
 import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -20,13 +20,13 @@ import org.springframework.web.bind.annotation.RestController;
 public class AuthController {
 
     @Autowired
-    private UserService userService;
+    private AuthService authService;
 
     @PostMapping("/register")
     public ResponseEntity<?> registerUser(@RequestBody UserRequest userRequest, HttpServletRequest request) throws Exception {
        String url = CommonUtil.getUrl(request);
 
-        Boolean register = userService.register(userRequest, url);
+        Boolean register = authService.register(userRequest, url);
 
         if (register) {
             return CommonUtil.createBuildResponseMessage("User registered successfully", HttpStatus.CREATED);
@@ -37,7 +37,7 @@ public class AuthController {
 
     @PostMapping("/login")
     public ResponseEntity<?> login(@RequestBody LoginRequest loginRequest) throws Exception {
-        LoginResponse loginResponse = userService.login(loginRequest);
+        LoginResponse loginResponse = authService.login(loginRequest);
 
         if (ObjectUtils.isEmpty(loginResponse)) {
             return CommonUtil.createErrorResponseMessage("Invalid credentials. Login failed", HttpStatus.BAD_REQUEST);
