+++ src/main/java/com/example/enotes/controller/UserController.java
@@ -1,15 +1,15 @@
 package com.example.enotes.controller;
 
+import com.example.enotes.dto.PasswordChangeRequest;
 import com.example.enotes.dto.UserResponse;
 import com.example.enotes.entity.User;
+import com.example.enotes.service.UserService;
 import com.example.enotes.util.CommonUtil;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequestMapping("/api/v1/user")
@@ -18,11 +18,21 @@ public class UserController {
     @Autowired
     private ModelMapper mapper;
 
+    @Autowired
+    private UserService userService;
+
     @GetMapping("/profile")
     public ResponseEntity<?> getProfile() {
         User loggedInUser = CommonUtil.getLoggedInUser();
         UserResponse userResponse = mapper.map(loggedInUser, UserResponse.class);
 
         return CommonUtil.createBuildResponse(userResponse, HttpStatus.OK);
+    }
+
+    @PostMapping("/change-password")
+    public ResponseEntity<?> changePassword(@RequestBody PasswordChangeRequest passwordChangeRequest) {
+        userService.changePassword(passwordChangeRequest);
+
+        return CommonUtil.createBuildResponseMessage("Password Changed Successfully", HttpStatus.OK);
     }
 }
