+++ src/main/java/com/example/enotes/dto/LoginResponse.java
@@ -7,7 +7,7 @@ import lombok.Data;
 @Builder
 public class LoginResponse {
 
-    private UserRequest user;
+    private UserResponse user;
 
     private String token;
 }
