+++ src/main/java/com/example/enotes/dto/PasswordChangeRequest.java
@@ -1,0 +1,15 @@
+package com.example.enotes.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class PasswordChangeRequest {
+
+    private String oldPassword;
+
+    private String newPassword;
+}
