+++ src/main/java/com/example/enotes/service/AuthService.java
@@ -1,0 +1,11 @@
+package com.example.enotes.service;
+
+import com.example.enotes.dto.LoginRequest;
+import com.example.enotes.dto.LoginResponse;
+import com.example.enotes.dto.UserRequest;
+
+public interface AuthService {
+    public Boolean register(UserRequest userRequest, String url) throws Exception;
+
+    LoginResponse login(LoginRequest loginRequest);
+}
